# server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is the documentation api of the fullstack-dev-assesment"
  version: "1.0.0"
  title: "Campaign API"
  contact:
    email: "mrmello92@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:9080"
tags:
- name: "campaign"
  description: "Ad Campaigns"
schemes:
- "http"
paths:
  /campaigns:
    get:
      tags:
      - "campaign"
      summary: "Lists all campaigns"
      operationId: "listAllCampaings"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/campaign"
        400:
          description: "Invalid request"
        500:
          description: "Internal error"
  /campaigns/{id}:
    get:
      tags:
      - "campaign"
      summary: "Find campaign by ID"
      description: "Returns a single campaign"
      operationId: "campaignDetail"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of campaign to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/campaign"
        400:
          description: "Invalid ID supplied | ID not found"
        500:
          description: "Internal error"
definitions:
  campaign:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      goal:
        type: "string"
      total_budget:
        type: "integer"
        format: "int32"
      status:
        type: "string"
      platforms:
        $ref: "#/definitions/platforms"
  platforms:
    type: "object"
    properties:
      facebook:
        $ref: "#/definitions/facebook"
      instagram:
        $ref: "#/definitions/instagram"
      google:
        $ref: "#/definitions/google"
  facebook:
    type: "object"
    properties:
      status:
        type: "string"
      total_budget:
        type: "integer"
        format: "int32"
      remaining_budget:
        type: "integer"
        format: "int32"
      start_date:
        type: "integer"
        format: "int64"
      end_date:
        type: "integer"
        format: "int64"
      target_audiance:
        $ref: "#/definitions/target_audiance"
      creatives:
        $ref: "#/definitions/creatives"
      insights:
        $ref: "#/definitions/insights"
  instagram:
    type: "object"
    properties:
      status:
        type: "string"
      total_budget:
        type: "integer"
        format: "int32"
      remaining_budget:
        type: "integer"
        format: "int32"
      start_date:
        type: "integer"
        format: "int64"
      end_date:
        type: "integer"
        format: "int64"
      target_audiance:
        $ref: "#/definitions/target_audiance"
      creatives:
        $ref: "#/definitions/creatives"
      insights:
        $ref: "#/definitions/insights"
  google:
    type: "object"
    properties:
      status:
        type: "string"
      total_budget:
        type: "integer"
        format: "int32"
      remaining_budget:
        type: "integer"
        format: "int32"
      start_date:
        type: "integer"
        format: "int64"
      end_date:
        type: "integer"
        format: "int64"
      target_audiance:
        $ref: "#/definitions/target_audiance"
      creatives:
        $ref: "#/definitions/creatives"
      insights:
        $ref: "#/definitions/insights"
  target_audiance:
    type: "object"
    properties:
      languages:
        type: "array"
        items:
          type: string
      genders:
        type: "array"
        items:
          type: string
      age_range:
        type: "array"
        items:
          type: string
      locations:
        type: "array"
        items:
          type: string
      interests:
        type: "array"
        items:
          type: string
  creatives:
    type: "object"
    properties:
      header:
        type: "string"
      description:
        type: "string"
      url:
        type: "string"
      image:
        type: "string"
  insights:
    type: "object"
    properties:
      impressions:
        type: "number"
      clicks:
        type: "number"
      nanos_score:
        type: "number"
      cost_per_click:
        type: "number"
      click_through_rate:
        type: "number"
      advanced_kpi_1:
        type: "number"
      advanced_kpi_2:
        type: "number"